requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
